+++ src/app/constants.js
@@ -45,19 +45,77 @@ export const FileType = {
  */
 export const CoordinationType = {
   DATASET: 'dataset',
+
+  // Observation type (e.g., cell)
+  OBS_TYPE: 'obsType',
+  OBS_FILTER: 'obsFilter',
+  OBS_HIGHLIGHT: 'obsHighlight',
+  OBS_SELECTION: 'obsSelection',
+  OBS_SET_SELECTION: 'obsSetSelection',
+  OBS_SET_HIGHLIGHT: 'obsSetHighlight',
+  OBS_SET_COLOR: 'obsSetColor',
+  OBS_COLOR_ENCODING: 'obsColorEncoding', // values: obsSetSelection, featureSelection, obsColor
+
+  // Sub-observation type (e.g., molecule)
+  SUB_OBS_TYPE: 'subObsType',
+  SUB_OBS_FILTER: 'subObsFilter',
+  SUB_OBS_HIGHLIGHT: 'subObsHighlight',
+  SUB_OBS_SELECTION: 'subObsSelection',
+  SUB_OBS_SET_SELECTION: 'subObsSetSelection',
+  SUB_OBS_SET_HIGHLIGHT: 'subObsSetHighlight',
+  SUB_OBS_SET_COLOR: 'subObsSetColor',
+  SUB_OBS_COLOR_ENCODING: 'subObsColorEncoding', // values: subObsSetSelection, subFeatureSelection, subObsColor
+
+  // Feature type (e.g., gene)
+  FEATURE_TYPE: 'featureType',
+  FEATURE_FILTER: 'featureFilter',
+  FEATURE_HIGHLIGHT: 'featureHighlight',
+  FEATURE_SELECTION: 'featureSelection',
+  FEATURE_SET_SELECTION: 'featureSetSelection',
+  FEATURE_SET_HIGHLIGHT: 'featureSetHighlight',
+  FEATURE_SET_COLOR: 'featureSetColor',
+
+  // Sub-feature type (e.g., transcript)
+  SUB_FEATURE_TYPE: 'subFeatureType',
+  SUB_FEATURE_FILTER: 'subFeatureFilter',
+  SUB_FEATURE_HIGHLIGHT: 'subFeatureHighlight',
+  SUB_FEATURE_SELECTION: 'subFeatureSelection',
+  SUB_FEATURE_SET_SELECTION: 'subFeatureSetSelection',
+  SUB_FEATURE_SET_HIGHLIGHT: 'subFeatureSetHighlight',
+  SUB_FEATURE_SET_COLOR: 'subFeatureSetColor',
+
+  // Feature value type (e.g., expression)
+  FEATURE_VALUE_TYPE: 'featureValueType',
+  FEATURE_VALUE_COLORMAP: 'featureValueColormap',
+  FEATURE_VALUE_COLORMAP_RANGE: 'featureValueColormapRange',
+  FEATURE_VALUE_TRANSFORM: 'featureValueTransform',
+
+  // Sub-feature value type (e.g., intensity)
+  SUB_FEATURE_VALUE_TYPE: 'subFeatureValueType',
+  SUB_FEATURE_VALUE_COLORMAP: 'subFeatureValueColormap',
+  SUB_FEATURE_VALUE_COLORMAP_RANGE: 'subFeatureValueColormapRange',
+  SUB_FEATURE_VALUE_TRANSFORM: 'subFeatureValueTransform',
+
   EMBEDDING_TYPE: 'embeddingType',
   EMBEDDING_ZOOM: 'embeddingZoom',
   EMBEDDING_ROTATION: 'embeddingRotation',
   EMBEDDING_TARGET_X: 'embeddingTargetX',
   EMBEDDING_TARGET_Y: 'embeddingTargetY',
   EMBEDDING_TARGET_Z: 'embeddingTargetZ',
-  EMBEDDING_CELL_SET_POLYGONS_VISIBLE: 'embeddingCellSetPolygonsVisible',
-  EMBEDDING_CELL_SET_LABELS_VISIBLE: 'embeddingCellSetLabelsVisible',
-  EMBEDDING_CELL_SET_LABEL_SIZE: 'embeddingCellSetLabelSize',
-  EMBEDDING_CELL_RADIUS: 'embeddingCellRadius',
-  EMBEDDING_CELL_RADIUS_MODE: 'embeddingCellRadiusMode',
-  EMBEDDING_CELL_OPACITY: 'embeddingCellOpacity',
-  EMBEDDING_CELL_OPACITY_MODE: 'embeddingCellOpacityMode',
+  EMBEDDING_OBS_SET_POLYGONS_VISIBLE: 'embeddingObsSetPolygonsVisible',
+  EMBEDDING_OBS_SET_LABELS_VISIBLE: 'embeddingObsSetLabelsVisible',
+  EMBEDDING_OBS_SET_LABEL_SIZE: 'embeddingObsSetLabelSize',
+  EMBEDDING_OBS_RADIUS: 'embeddingObsRadius',
+  EMBEDDING_OBS_RADIUS_MODE: 'embeddingObsRadiusMode',
+  EMBEDDING_OBS_OPACITY: 'embeddingObsOpacity',
+  EMBEDDING_OBS_OPACITY_MODE: 'embeddingObsOpacityMode',
+  EMBEDDING_CELL_SET_POLYGONS_VISIBLE: 'embeddingCellSetPolygonsVisible', // deprecate
+  EMBEDDING_CELL_SET_LABELS_VISIBLE: 'embeddingCellSetLabelsVisible', // deprecate
+  EMBEDDING_CELL_SET_LABEL_SIZE: 'embeddingCellSetLabelSize', // deprecate
+  EMBEDDING_CELL_RADIUS: 'embeddingCellRadius', // deprecate
+  EMBEDDING_CELL_RADIUS_MODE: 'embeddingCellRadiusMode', // deprecate
+  EMBEDDING_CELL_OPACITY: 'embeddingCellOpacity', // deprecate
+  EMBEDDING_CELL_OPACITY_MODE: 'embeddingCellOpacityMode', // deprecate
   SPATIAL_ZOOM: 'spatialZoom',
   SPATIAL_ROTATION: 'spatialRotation',
   SPATIAL_TARGET_X: 'spatialTargetX',
@@ -73,26 +131,32 @@ export const CoordinationType = {
   HEATMAP_ZOOM_Y: 'heatmapZoomY',
   HEATMAP_TARGET_X: 'heatmapTargetX',
   HEATMAP_TARGET_Y: 'heatmapTargetY',
-  CELL_FILTER: 'cellFilter',
-  CELL_HIGHLIGHT: 'cellHighlight',
-  CELL_SET_SELECTION: 'cellSetSelection',
-  CELL_SET_HIGHLIGHT: 'cellSetHighlight',
-  CELL_SET_COLOR: 'cellSetColor',
-  GENE_FILTER: 'geneFilter',
-  GENE_HIGHLIGHT: 'geneHighlight',
-  GENE_SELECTION: 'geneSelection',
-  GENE_EXPRESSION_COLORMAP: 'geneExpressionColormap',
-  GENE_EXPRESSION_TRANSFORM: 'geneExpressionTransform',
-  GENE_EXPRESSION_COLORMAP_RANGE: 'geneExpressionColormapRange',
-  CELL_COLOR_ENCODING: 'cellColorEncoding',
+  CELL_FILTER: 'cellFilter', // deprecate
+  CELL_HIGHLIGHT: 'cellHighlight', // deprecate
+  CELL_SET_SELECTION: 'cellSetSelection', // deprecate
+  CELL_SET_HIGHLIGHT: 'cellSetHighlight', // deprecate
+  CELL_SET_COLOR: 'cellSetColor', // deprecate
+  GENE_FILTER: 'geneFilter', // deprecate
+  GENE_HIGHLIGHT: 'geneHighlight', // deprecate
+  GENE_SELECTION: 'geneSelection', // deprecate
+  GENE_EXPRESSION_COLORMAP: 'geneExpressionColormap', // deprecate
+  GENE_EXPRESSION_TRANSFORM: 'geneExpressionTransform', // deprecate
+  GENE_EXPRESSION_COLORMAP_RANGE: 'geneExpressionColormapRange', // deprecate
+  CELL_COLOR_ENCODING: 'cellColorEncoding', // deprecate
   SPATIAL_RASTER_LAYERS: 'spatialRasterLayers',
-  SPATIAL_CELLS_LAYER: 'spatialCellsLayer',
-  SPATIAL_MOLECULES_LAYER: 'spatialMoleculesLayer',
+  SPATIAL_CELLS_LAYER: 'spatialCellsLayer', // deprecate
+  SPATIAL_MOLECULES_LAYER: 'spatialMoleculesLayer', // deprecate
+  SPATIAL_OBS_LAYER: 'spatialObsLayer',
+  SPATIAL_SUB_OBS_LAYER: 'spatialSubObsLayer',
   SPATIAL_NEIGHBORHOODS_LAYER: 'spatialNeighborhoodsLayer',
   GENOMIC_ZOOM_X: 'genomicZoomX',
   GENOMIC_ZOOM_Y: 'genomicZoomY',
   GENOMIC_TARGET_X: 'genomicTargetX',
   GENOMIC_TARGET_Y: 'genomicTargetY',
-  ADDITIONAL_CELL_SETS: 'additionalCellSets',
-  MOLECULE_HIGHLIGHT: 'moleculeHighlight',
+  ADDITIONAL_OBS_SETS: 'additionalObsSets',
+  ADDITIONAL_SUB_OBS_SETS: 'additionalSubObsSets',
+  ADDITIONAL_FEATURE_SETS: 'additionalFeatureSets',
+  ADDITIONAL_SUB_FEATURE_SETS: 'additionalSubFeatureSets',
+  ADDITIONAL_CELL_SETS: 'additionalCellSets', // deprecate
+  MOLECULE_HIGHLIGHT: 'moleculeHighlight', // deprecate
 };